// dashboard/src/client/models/create-user-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Scheduler Public API (v1)
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface CreateUserRequest {
    'username': string;
    'displayName': string;
    'password': string;
    /**
     * Optional initial roles to assign to the user.
     */
    'roles'?: Array<string>;
}