// dashboard/src/client/models/user-summary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Scheduler Public API (v1)
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface UserSummary {
    'userId': string;
    'username': string;
    'displayName': string;
    'roles': Array<string>;
    /**
     * Whether the account is enabled.
     */
    'isActive': boolean;
}